Extract tierEnvelopeFactory
We extract tierEnvelopeFactory into its own file, as the pervious method of inclusion in factories/index.ts seems to potentially confuse our test coverage checker, causing it to report that this file has only 50% coverage
