+++ server/testutils/factories/index.ts
@@ -47,9 +47,10 @@ import premisesSummaryFactory from './premisesSummary'
 import prisonCaseNotesFactory from './prisonCaseNotes'
 import probationRegionFactory from './probationRegion'
 import referenceDataFactory from './referenceData'
-import risksFactory, { tierEnvelopeFactory } from './risks'
+import risksFactory from './risks'
 import roomFactory from './room'
 import staffMemberFactory from './staffMember'
+import tierEnvelopeFactory from './tierEnvelopeFactory'
 import turnaroundFactory from './turnaround'
 import updateLostBedFactory from './updateLostBed'
 import updatePremisesFactory from './updatePremises'
