+++ server/testutils/factories/risks.ts
@@ -1,12 +1,12 @@
 import { faker } from '@faker-js/faker/locale/en_GB'
 import { Factory } from 'fishery'
 
-import type { PersonRisks, RiskEnvelopeStatus } from '@approved-premises/api'
-import { RiskLevel, RiskTierLevel, TierLetter, TierNumber } from '@approved-premises/ui'
+import type { PersonRisks } from '@approved-premises/api'
+import { RiskLevel } from '@approved-premises/ui'
 import { DateFormats } from '../../utils/dateUtils'
+import tierEnvelopeFactory, { riskEnvelopeStatuses } from './tierEnvelopeFactory'
 
 const riskLevels: Array<RiskLevel> = ['Low', 'Medium', 'High', 'Very High']
-const riskEnvelopeStatuses: Array<RiskEnvelopeStatus> = ['retrieved', 'not_found', 'error']
 
 export default Factory.define<PersonRisks>(() => ({
   crn: `C${faker.number.int({ min: 100000, max: 999999 })}`,
@@ -48,16 +48,3 @@ const flagsFactory = Factory.define<PersonRisks['flags']>(() => {
     ]),
   }
 })
-
-const lettersFactory: () => TierLetter = () => faker.helpers.arrayElement<TierLetter>(['A', 'B', 'C', 'D'])
-const numbersFactory: () => TierNumber = () => faker.helpers.arrayElement<TierNumber>(['1', '2', '3', '4'])
-
-export const riskTierLevel: RiskTierLevel = `${lettersFactory()}${numbersFactory()}`
-
-export const tierEnvelopeFactory = Factory.define<PersonRisks['tier']>(() => ({
-  status: faker.helpers.arrayElement(riskEnvelopeStatuses),
-  value: {
-    level: riskTierLevel,
-    lastUpdated: DateFormats.dateObjToIsoDate(faker.date.past()),
-  },
-}))
